cleaner tweet fetching
