+++ package.json
@@ -12,6 +12,7 @@
 		"chalk": "^2.4.1",
 		"dotenv": "^5.0.1",
 		"express": "^4.16.3",
+		"jest": "^23.0.1",
 		"js-yaml": "^3.11.0",
 		"puppeteer": "^1.4.0",
 		"twitter": "^1.7.1",
