--- txt/_cache.js
@@ -1,424 +1,0 @@
-module.exports = [
-	{
-		created_at: 'Thu May 24 05:24:55 +0000 2018',
-		id: 999521606325145600,
-		id_str: '999521606325145600',
-		text:
-			'John Lee Hooker: Boom boom boom boom (guitar riff)\nme: hm pretty good. What else you got\nJohn Lee Hooker: Hau hau hau hau\nme: Fuck this .',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1884,
-		favorite_count: 13351,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Wed May 23 22:09:40 +0000 2018',
-		id: 999412071187009500,
-		id_str: '999412071187009537',
-		text:
-			'#worldturtleday these beasts are simply armoured rats ,always and forever',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 2218,
-		favorite_count: 9861,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Wed May 23 22:05:17 +0000 2018',
-		id: 999410966365392900,
-		id_str: '999410966365392896',
-		text:
-			'babies cant smile. its a biological reflex from the brain stem. its fake. they dont know how to smile',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 2031,
-		favorite_count: 13551,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Wed May 23 21:57:26 +0000 2018',
-		id: 999408992710463500,
-		id_str: '999408992710463488',
-		text:
-			'Ok bitch write "Telomere Mindfulnness" on a piece of paper. Go to Golds gym on 23rd, hand the paper to "Eddie" &amp; tell him youre ready toDie',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 845,
-		favorite_count: 4922,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Wed May 23 21:43:31 +0000 2018',
-		id: 999405491720802300,
-		id_str: '999405491720802305',
-		text:
-			'Time to show these douscebags my real power (throws a big tree branch at a guys car)',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1393,
-		favorite_count: 8349,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Tue May 22 21:19:37 +0000 2018',
-		id: 999037087851667500,
-		id_str: '999037087851667463',
-		text:
-			'seeing the cops in the rear view, telling everyone to "stay cool" and cutting hard into the guardrail to flip my car on purpose',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com/download/android" rel="nofollow">Twitter for Android</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 4235,
-		favorite_count: 21833,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Mon May 21 21:50:38 +0000 2018',
-		id: 998682506298888200,
-		id_str: '998682506298888192',
-		text: 'https://t.co/qCucXEEUH5',
-		truncated: false,
-		entities: [Object],
-		extended_entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com/download/android" rel="nofollow">Twitter for Android</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 913,
-		favorite_count: 8268,
-		favorited: false,
-		retweeted: false,
-		possibly_sensitive: false,
-		lang: 'und',
-	},
-	{
-		created_at: 'Mon May 21 20:59:53 +0000 2018',
-		id: 998669733460435000,
-		id_str: '998669733460434944',
-		text:
-			'(after getting split in half by a 16-inch bullet fired from a boat) waiter theres a fly in my Soup',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1145,
-		favorite_count: 8306,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Mon May 21 20:27:09 +0000 2018',
-		id: 998661497910939600,
-		id_str: '998661497910939648',
-		text:
-			'ME: What do you mean the plane departed. The clock says its only 8:25.\nTICKET AGENT: Sir Thats a clock\nME: I know',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1165,
-		favorite_count: 9126,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Mon May 21 18:41:31 +0000 2018',
-		id: 998634912981966800,
-		id_str: '998634912981966848',
-		text:
-			'james bonds jet black diapers that cost $200 each. suit jacket that has Jame bonds signature drawn on the pocket. james bond engagement ring',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 431,
-		favorite_count: 3972,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Mon May 21 18:33:38 +0000 2018',
-		id: 998632929587880000,
-		id_str: '998632929587879937',
-		text:
-			'the guy who hooks me up with Raw Milk locked his account. too many people posted "come to brazil" at him so i guess ill starve to death now',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1696,
-		favorite_count: 11810,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Mon May 21 18:22:38 +0000 2018',
-		id: 998630162160812000,
-		id_str: '998630162160812033',
-		text:
-			'before you threaten to "Pimp slap" me, please remember that i am a police officer of the law, on at least four separate timeline\'s',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 2050,
-		favorite_count: 12235,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Sun May 20 03:47:51 +0000 2018',
-		id: 998047627395719200,
-		id_str: '998047627395719168',
-		text:
-			'"octo mom" will never be allowed to join mensa or the fortune 500.  just saying the damn truth like it is',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1159,
-		favorite_count: 11766,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Fri May 18 22:35:47 +0000 2018',
-		id: 997606704853520400,
-		id_str: '997606704853520384',
-		text:
-			'-tweeting about stuff thats off-topic is "Random". never do this.\n-tweeting about current event stuff is "Topical". and this is also a crime',
-		truncated: false,
-		entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 3628,
-		favorite_count: 19598,
-		favorited: false,
-		retweeted: false,
-		lang: 'en',
-	},
-	{
-		created_at: 'Thu May 17 11:34:15 +0000 2018',
-		id: 997077836358811600,
-		id_str: '997077836358811649',
-		text: 'https://t.co/nwBs9p35Oi',
-		truncated: false,
-		entities: [Object],
-		extended_entities: [Object],
-		metadata: [Object],
-		source:
-			'<a href="http://twitter.com" rel="nofollow">Twitter Web Client</a>',
-		in_reply_to_status_id: null,
-		in_reply_to_status_id_str: null,
-		in_reply_to_user_id: null,
-		in_reply_to_user_id_str: null,
-		in_reply_to_screen_name: null,
-		user: [Object],
-		geo: null,
-		coordinates: null,
-		place: null,
-		contributors: null,
-		is_quote_status: false,
-		retweet_count: 1053,
-		favorite_count: 10741,
-		favorited: false,
-		retweeted: false,
-		possibly_sensitive: false,
-		lang: 'und',
-	},
-];
