+++ txt/cache.json
@@ -1,0 +1,56 @@
+[
+	{
+		"text":
+			"John Lee Hooker: Boom boom boom boom (guitar riff)\nme: hm pretty good. What else you got\nJohn Lee Hooker: Hau hau hau hau\nme: Fuck this ."
+	},
+	{
+		"text":
+			"#worldturtleday these beasts are simply armoured rats ,always and forever"
+	},
+	{
+		"text":
+			"babies cant smile. its a biological reflex from the brain stem. its fake. they dont know how to smile"
+	},
+	{
+		"text":
+			"Ok bitch write \"Telomere Mindfulnness\" on a piece of paper. Go to Golds gym on 23rd, hand the paper to \"Eddie\" &amp; tell him youre ready toDie"
+	},
+	{
+		"text":
+			"Time to show these douscebags my real power (throws a big tree branch at a guys car)"
+	},
+	{
+		"text":
+			"seeing the cops in the rear view, telling everyone to \"stay cool\" and cutting hard into the guardrail to flip my car on purpose"
+	},
+	{ "text": "https://t.co/qCucXEEUH5" },
+	{
+		"text":
+			"(after getting split in half by a 16-inch bullet fired from a boat) waiter theres a fly in my Soup"
+	},
+	{
+		"text":
+			"ME: What do you mean the plane departed. The clock says its only 8:25.\nTICKET AGENT: Sir Thats a clock\nME: I know"
+	},
+	{
+		"text":
+			"james bonds jet black diapers that cost $200 each. suit jacket that has Jame bonds signature drawn on the pocket. james bond engagement ring"
+	},
+	{
+		"text":
+			"the guy who hooks me up with Raw Milk locked his account. too many people posted \"come to brazil\" at him so i guess ill starve to death now"
+	},
+	{
+		"text":
+			"before you threaten to \"Pimp slap\" me, please remember that i am a police officer of the law, on at least four separate timeline's"
+	},
+	{
+		"text":
+			"\"octo mom\" will never be allowed to join mensa or the fortune 500.  just saying the damn truth like it is"
+	},
+	{
+		"text":
+			"-tweeting about stuff thats off-topic is \"Random\". never do this.\n-tweeting about current event stuff is \"Topical\". and this is also a crime"
+	},
+	{ "text": "https://t.co/nwBs9p35Oi" }
+]
